Introduce factory for creation of hitcount threshold checkers
